+++ packages/components/autocomplete/src/multi-autocomplete.tsx
@@ -310,7 +310,7 @@ const MultiAutocompleteField = forwardRef<MultiAutocompleteFieldProps, "input">(
     }, [label, component, value, onChange, isOpen, inputRef, separator])
 
     const css: CSSUIObject = {
-      paddingEnd: "2rem",
+      pe: "2rem",
       h,
       minH,
       display: "flex",