+++ packages/components/resizable/src/index.ts
@@ -1,0 +1,10 @@
+export { Resizable } from "./resizable"
+export type { ResizableProps } from "./resizable"
+export { ResizableItem } from "./resizable-item"
+export type { ResizableItemProps } from "./resizable-item"
+export { ResizableTrigger, ResizableTriggerIcon } from "./resizable-trigger"
+export type {
+  ResizableTriggerProps,
+  ResizableTriggerIconProps,
+} from "./resizable-trigger"
+export type { ResizableStorage, ResizableItemControl } from "./use-resizable"