+++ packages/components/calendar/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @yamada-ui/calendar
 
+## 1.1.5
+
+### Patch Changes
+
+- [#622](https://github.com/hirotomoyamada/yamada-ui/pull/622) [`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf) Thanks [@hirotomoyamada](https://github.com/hirotomoyamada)! - Formatted component internal styles.
+
+- Updated dependencies [[`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf), [`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf), [`b678019`](https://github.com/hirotomoyamada/yamada-ui/commit/b6780198dc3441eac04bf1f7fd9210c726f32edb), [`73b2111`](https://github.com/hirotomoyamada/yamada-ui/commit/73b211195ffa3dd8a2e10897ccf103549b10906c), [`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf)]:
+  - @yamada-ui/utils@1.0.4
+  - @yamada-ui/core@1.3.0
+  - @yamada-ui/button@1.0.10
+  - @yamada-ui/form-control@1.0.11
+  - @yamada-ui/icon@1.0.9
+  - @yamada-ui/popover@1.0.9
+  - @yamada-ui/portal@1.0.4
+  - @yamada-ui/use-clickable@1.0.4
+  - @yamada-ui/use-controllable-state@1.0.4
+  - @yamada-ui/use-outside-click@1.0.4
+
 ## 1.1.4
 
 ### Patch Changes