+++ packages/components/tooltip/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @yamada-ui/tooltip
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf), [`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf), [`b678019`](https://github.com/hirotomoyamada/yamada-ui/commit/b6780198dc3441eac04bf1f7fd9210c726f32edb), [`73b2111`](https://github.com/hirotomoyamada/yamada-ui/commit/73b211195ffa3dd8a2e10897ccf103549b10906c), [`021f818`](https://github.com/hirotomoyamada/yamada-ui/commit/021f818babd2d04a8163807fa6fe8122b60738bf)]:
+  - @yamada-ui/utils@1.0.4
+  - @yamada-ui/core@1.3.0
+  - @yamada-ui/motion@1.0.9
+  - @yamada-ui/portal@1.0.4
+  - @yamada-ui/transitions@1.0.9
+  - @yamada-ui/use-disclosure@1.0.4
+  - @yamada-ui/use-event-listener@1.0.4
+  - @yamada-ui/use-popper@1.0.9
+
 ## 1.0.9
 
 ### Patch Changes